requirements.txt
aniso8601==9.0.1
asgiref==3.7.1
astroid==2.15.4
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dill==0.3.6
distlib==0.3.6
Django==4.2.1
django-crispy-forms==2.0
djangorestframework==3.14.0
filelock==3.11.0
Flask==2.2.3
Flask-MySQLdb==1.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.2.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mysqlclient==2.1.1
numpy==1.24.1
Pillow==9.5.0
platformdirs==3.2.0
psycopg2==2.9.6
pycodestyle==2.10.0
pylint==2.17.4
PyMySQL==1.0.3
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.11
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.11.8
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.21.0
Werkzeug==2.2.3
wrapt==1.15.0
zipp==3.15.0
